Improved testing framework
The testing framework is now less hackish. Many tests have been rewritten with mockups of only the
relevant parts of the simulator.
Still have to rework:
- msg_q test
- LP tests
- GVT tests

Some visibility tests are currently disabled, because they refer to a shared object
implementation, but we're only shipping a static library.
