--- test/integration/model/output_256.c
@@ -1,268 +1,0 @@
-/**
- * @file test/integration/model/output_256.c
- *
- * @brief Correct output of the model used to verify correctness (with 256 LPs)
- *
- * SPDX-FileCopyrightText: 2008-2021 HPDCS Group <rootsim@googlegroups.com>
- * SPDX-License-Identifier: GPL-3.0-only
- */
-#include <integration/model/application.h>
-
-const uint32_t model_expected_output[256] = {
-	1954882895,
-	499914811,
-	2172615350,
-	855680666,
-	4078967830,
-	1967356563,
-	1244932102,
-	3500001537,
-	2369782773,
-	2839972308,
-	3064712116,
-	2808073134,
-	895501456,
-	660702574,
-	3363720750,
-	2987230612,
-	3244500016,
-	3683442101,
-	878909684,
-	2124109236,
-	2616537132,
-	73871941,
-	2283315819,
-	2585581244,
-	1633673897,
-	926438166,
-	1744373315,
-	1244361757,
-	3659793495,
-	1121184652,
-	2812625974,
-	2299640512,
-	3901424024,
-	3653269204,
-	2236798247,
-	3025496743,
-	499842957,
-	3915660410,
-	2086850946,
-	125001660,
-	910493557,
-	3652986087,
-	591923086,
-	4003950471,
-	2014028941,
-	3714477812,
-	1953882731,
-	3663001492,
-	1361201354,
-	838409246,
-	3031574938,
-	2020575815,
-	305108375,
-	2378084545,
-	2838187189,
-	1989468954,
-	4163057611,
-	155716475,
-	2842892188,
-	2523094577,
-	2672304035,
-	2106435401,
-	1900383561,
-	2045511042,
-	2716752688,
-	1749598857,
-	2638651666,
-	3824997262,
-	1230689068,
-	3774698042,
-	3749704855,
-	446768214,
-	958584014,
-	1741478016,
-	2358822956,
-	601539315,
-	3166014770,
-	1305013943,
-	108336892,
-	626874566,
-	2608477477,
-	4238243585,
-	1512922685,
-	1681319003,
-	1258695965,
-	595874774,
-	1713022936,
-	171166593,
-	3260203371,
-	3273802937,
-	4042122267,
-	303434850,
-	2959432150,
-	4085181192,
-	1747957427,
-	265285831,
-	3662825970,
-	4200286206,
-	524931647,
-	1105723880,
-	4055358905,
-	341691096,
-	1102271707,
-	1348974794,
-	2331167931,
-	3787302218,
-	1252206999,
-	2138124713,
-	1485749551,
-	1479135728,
-	862367699,
-	2046180142,
-	539599401,
-	3637780453,
-	2575688978,
-	2104233330,
-	3228832648,
-	1309131876,
-	2774537706,
-	4235410709,
-	1350279773,
-	2149757407,
-	2997196212,
-	1088821636,
-	821875211,
-	3461491312,
-	3175104820,
-	2670625452,
-	94578962,
-	331121276,
-	843695503,
-	1196885102,
-	2359473962,
-	2437515188,
-	2759110588,
-	4245129350,
-	4065830373,
-	2298067264,
-	2446505925,
-	372988225,
-	1476378066,
-	2840168646,
-	493888536,
-	186099399,
-	398123805,
-	2220966016,
-	2648475982,
-	1850764574,
-	2489606902,
-	903912017,
-	1940627452,
-	3117601413,
-	3655419710,
-	819516011,
-	3525510125,
-	1491790265,
-	2435157392,
-	4118825661,
-	3900814958,
-	2421024176,
-	4121017118,
-	3996807550,
-	2988422848,
-	3176612759,
-	223771054,
-	104508114,
-	4082962868,
-	1715413243,
-	3581130575,
-	940701070,
-	907255229,
-	4225184216,
-	2467675262,
-	3061676384,
-	2229033610,
-	3407087143,
-	954634198,
-	2988531099,
-	31552366,
-	4037116824,
-	3981634631,
-	2181493081,
-	2443323686,
-	907224002,
-	3476965318,
-	561849374,
-	889553565,
-	2449620640,
-	1335476373,
-	3421917346,
-	1720110291,
-	2865322561,
-	190563775,
-	3089656667,
-	4069361454,
-	1670178052,
-	399696533,
-	1976893687,
-	548593047,
-	657213459,
-	166817696,
-	893758812,
-	1023297306,
-	3554592332,
-	776299921,
-	993599013,
-	2597361605,
-	3794025879,
-	1543603020,
-	2578300288,
-	823909038,
-	654320367,
-	2910565513,
-	3956772967,
-	3529191177,
-	2653335478,
-	2143878115,
-	1616242440,
-	1523897596,
-	2034594492,
-	2357314357,
-	1517155650,
-	358539191,
-	425604091,
-	1304996654,
-	2511252571,
-	799964978,
-	332547154,
-	3159264316,
-	1919035425,
-	655759176,
-	1530893137,
-	2098042536,
-	745654059,
-	2286662715,
-	2244909142,
-	3915019597,
-	2299076255,
-	2491672189,
-	1659536272,
-	2359921760,
-	4258386414,
-	3447130890,
-	3618334635,
-	3924890068,
-	4245000590,
-	515223874,
-	2391163600,
-	3563634379,
-	1570670263,
-	3004255346,
-	3673560042,
-	1241581495,
-	100350193,
-	1935043821,
-	3213574777
-};
